* Fix a large number of warnings relating to signed/unsigned conversion and data loss

+++ gamed/src/Notifiers.cpp
@@ -26,13 +26,13 @@ void Game::notifyMovement(Object* o) {
    const std::vector<MovementVector>& waypoints = o->getWaypoints();
    MovementAns *answer = MovementAns::create(waypoints.size()*2);
    
-   for(int i = 0; i < waypoints.size(); i++) {
+   for(size_t i = 0; i < waypoints.size(); i++) {
       printf("     Vector %i, x: %f, y: %f\n", i, 2.0 * waypoints[i].x + MAP_WIDTH, 2.0 * waypoints[i].y + MAP_HEIGHT);
    }
    
    answer->nbUpdates = 1;
    answer->netId = o->getNetId();
-   for(int i = 0; i < waypoints.size(); i++) {
+   for(size_t i = 0; i < waypoints.size(); i++) {
       answer->getVector(i)->x = waypoints[i].x;
       answer->getVector(i)->y = waypoints[i].y;
    }
