+++ gamed/src/Object.cpp
@@ -35,7 +35,7 @@ void Object::setTarget(Target* target) {
 
 }
 
-void Object::Move(unsigned int diff) {
+void Object::Move(int64 diff) {
 
 	if(!target)
 	  return;
@@ -57,14 +57,14 @@ void Object::Move(unsigned int diff) {
 	}
 }
 
-void Object::update(unsigned int diff) {
+void Object::update(int64 diff) {
    Move(diff);
 }
 
 void Object::setWaypoints(const std::vector<MovementVector>& newWaypoints) {
    waypoints = newWaypoints;
    
-   setPosition(2.0 * waypoints[0].x + MAP_WIDTH, 2.0 * waypoints[0].y + MAP_HEIGHT);
+   setPosition(2.0f * waypoints[0].x + MAP_WIDTH, 2.0f * waypoints[0].y + MAP_HEIGHT);
    movementUpdated = true;
    if(waypoints.size() == 1) {
       setTarget(0);
